swap to new /post-list endpoint, update swr cache
