+++ components/news/article-list.js
@@ -1,4 +1,4 @@
-import { useNewsArticles } from "@/lib/strapi/newsGraphQL";
+import { useNewsArticles } from "@/lib/strapi/newsSWR";
 import { Box, Pagination, Skeleton, Stack, Typography } from "@mui/material"
 import { ArticlePreview } from "./article-preview";
 import { useMemo } from "react";
@@ -47,7 +47,7 @@ export const ArticleList = ({
 
   if (isLoading || !data) return <Stack><ArticleListSkeleton /></Stack>
 
-  const { articles, meta } = data;
+  const { results: articles, pagination } = data;
 
   if (Array.isArray(articles) && articles.length === 0) return <NoArticlesText />
 
@@ -65,7 +65,7 @@ export const ArticleList = ({
       ))}
     </Stack>
     
-    <Pagination count={meta.pagination?.pageCount} page={page} onChange={(_, p) => page === p ? null : setPage(p)} sx={{ alignSelf: 'center' }} />
+    <Pagination count={pagination?.pageCount} page={page} onChange={(_, p) => page === p ? null : setPage(p)} sx={{ alignSelf: 'center' }} />
   </Stack> 
 }
