+++ lib/strapi/index.js
@@ -7,4 +7,4 @@ export * from './groupsGraphQL'
 export * from './peopleGraphQL'
 export * from './fetchOurWorkTrayItems'
 export * from './newsAppearancesGraphQL'
-export * from './newsGraphQL'
+export * from './newsSWR'
