--- lib/strapi/newsGraphQL.js
@@ -1,125 +1,0 @@
-import { fetchStrapiGraphQL } from "./fetchStrapiGraphQL";
-import qs from 'qs';
-import useSWR from 'swr';
-
-export const useTags = () => {
-  const fetcher = (url) => fetch(url, { 
-    headers: {
-      'Accept': 'application/json',
-    },
-  })
-    .then((res) => {
-      if (!res.ok) throw new Error(`HTTP Error: ${res.status}`);
-      return res.json();
-    })
-    .then((data) => { 
-      if (data.error) throw new Error(`Strapi Error: ${data.error?.message ?? "Unknown error while fetching post tags"}`);
-      return data;
-    });
-
-  // dedupingInterval essentially sets cache expiration time to 1 hour
-  return useSWR('https://api.renci.org/api/all-post-tags', fetcher, { dedupingInterval: 1000 * 60 * 60 });
-}
-
-/**
- * SWR-enabled hook for fetch news article data
- *   @param {{
- *   filters: {
- *       researchGroups: string[],
- *       collaborations: string[],
- *       projects: string[],
- *       organizations: string[],
- *       people: string[],
- *       postTags: string[],
- *       freeSearch: string[],
- *       newsOrBlog: "news" | "blog" | undefined,
- *     },
- *     page?: number,
- *   }}
- */
-export const useNewsArticles = ({
-  filters: {
-    researchGroups,
-    collaborations,
-    projects,
-    organizations,
-    people,
-    postTags,
-    freeSearch,
-    newsOrBlog,
-  },
-  page = 1,
-}) => {
-  if ([researchGroups, collaborations, projects, organizations, people, freeSearch, postTags].some((t) => !Array.isArray(t))) {
-    throw new Error("One of the filters is not an array")
-  }
-
-  // make a list of all filters that need to be applied
-  const filtersList = [
-    ...researchGroups.map((rg) => ({ researchGroups: { slug: { $eq: rg }} })),
-    ...collaborations.map((c) => ({ collaborations: { slug: { $eq: c }} })),
-    ...projects.map((p) => ({ projects: { slug: { $eq: p }} })),
-    ...organizations.map((o) => ({ organizations: { slug: { $eq: o }} })),
-    ...people.map((p) => ({ people: { slug: { $eq: p }} })),
-    ...postTags.map((pt) => ({ tags: { name: { $eq: pt }} })),
-    ...freeSearch.map((fs) => ({
-      $or: [
-        {title: { $containsi: fs }},
-        {slug: { $containsi: fs }},
-      ]
-    })),
-    ...(newsOrBlog !== undefined ? [
-      { newsOrBlog: { $eq: newsOrBlog } }
-    ] : [])
-  ];
-
-  // if there is more than one filter, we need to boolean AND them together
-  // if there is just one, it needs to be spread to the top level of the `filters` obj
-  const filters = filtersList.length === 1
-    ? { ...filtersList[0] }
-    : { $and: filtersList }
-
-  const queryObj = {
-    sort: ['publishDate:desc'],
-    populate: "*",
-    fields: '*',
-    pagination: {
-      pageSize: 25,
-      page,
-    },
-    publicationState: 'live',
-    locale: ['en'],
-    filters,
-  };
-
-  const queryString = qs.stringify(queryObj, { encode: false });
-  const articlesUrl = `https://api.renci.org/api/posts?${queryString}`;
-
-  const fetcher = (url) => fetch(url, { 
-    headers: {
-      'Accept': 'application/json',
-    },
-  })
-    .then((res) => {
-      if (!res.ok) throw new Error(`HTTP Error: ${res.status}`);
-      return res.json();
-    })
-    .then((data) => { 
-      if (data.error) throw new Error(`Strapi Error: ${data.error?.message ?? "Unknown error while fetching articles"}`);
-      return ({
-        articles: data.entriesFormatted.map((article) => ({
-          ...article,
-          collaborations: article.collaborations.data.map(({ attributes }) => attributes),
-          people: article.people.data.map(({ attributes }) => ({ ...attributes, name: `${attributes.firstName} ${attributes.lastName}` })),
-          projects: article.projects.data.map(({ attributes }) => attributes),
-          organizations: article.organizations.data.map(({ attributes }) => attributes),
-          researchGroups: article.researchGroups.data.map(({ attributes }) => attributes),
-          tags: article.tags.data.map(({ attributes }) => attributes),
-        })),
-        meta: data.meta
-      });
-    });
-
-  // dedupingInterval essentially sets cache expiration time to 1 hour
-  return useSWR(articlesUrl, fetcher, { dedupingInterval: 1000 * 60 * 60 });
-}
